Changed shell module to be named mesh. Updated file names to all lowercase. Made necessary refactoring changes.
