+++ examples/write_abacus_solid_model.py
@@ -2,7 +2,7 @@ import pynumad as pynu
 import numpy as np
 import os
 
-from pynumad.shell.shell import getSolidMesh
+from pynumad.mesh.shell import getSolidMesh
 
 ## Read blade data from yaml file
 blade = pynu.objects.Blade.Blade()
