+++ src/pynumad/mesh/boundary2d.py
@@ -1,6 +1,6 @@
 import numpy as np
-import pynumad.shell.MeshTools as mt
-from pynumad.shell.Segment2DClass import *
+import pynumad.mesh.meshtools as mt
+from pynumad.mesh.segment2d import *
 
 
 class Boundary2D:
