+++ src/pynumad/mesh/mesh3d.py
@@ -1,6 +1,6 @@
 import numpy as np
 from scipy import interpolate
-from pynumad.shell.SpatialGridList3DClass import *
+from pynumad.mesh.spatialgridlist3d import *
 
 
 class Mesh3D:
