+++ src/pynumad/mesh/meshtools.py
@@ -1,7 +1,7 @@
 import numpy as np
 import plotly.graph_objects as go
-from pynumad.shell.SpatialGridList2DClass import *
-from pynumad.shell.SpatialGridList3DClass import *
+from pynumad.mesh.spatialgridlist2d import *
+from pynumad.mesh.spatialgridlist3d import *
 
 
 ## - Convert list of mesh objects into a single merged mesh, returning sets representing the elements/nodes from the original meshes
