+++ src/pynumad/tests/test_blade_io.py
@@ -1,5 +1,5 @@
 import unittest
-from os.path import abspath, dirname, join
+from os.path import join
 
 from pynumad.objects.Blade import Blade
 from pynumad.paths import DATA_PATH
@@ -14,10 +14,10 @@ class TestBladeIO(unittest.TestCase):
         self.yamlfile = join(test_data_dir, "blade_yamls", "BAR1_SNL_1_18_2021.yaml")
 
     def test_xlsx_blade(self):
-        xlsxblade = Blade(self.xlsxfile)
+        Blade(self.xlsxfile)
 
     def test_yaml_blade(self):
-        yamlblade = Blade(self.yamlfile)
+        Blade(self.yamlfile)
 
 
 if __name__ == "__main__":
